:sparkles::bulb: RegisterGifticonViewController 추가 (#35)
- RegisterGifticonTableVC 제거
