+++ Projects/App/Sources/Register/RegisterGifticonViewController.swift
@@ -1,0 +1,28 @@
+//
+//  RegisterGifticonViewController.swift
+//  GGiriGGiri
+//
+//  Created by 안상희 on 2022/07/09.
+//  Copyright © 2022 dvHuni. All rights reserved.
+//
+
+import UIKit
+
+final class RegisterGifticonViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        configure()
+    }
+    
+    private func configure() {
+        view.backgroundColor = .white
+        
+        configureNavigationBar()
+    }
+    
+    private func configureNavigationBar() {
+        navigationItem.title = "기프티콘 등록"
+    }
+}
